Create code.py
This code will ask for a probability threshold and then give you the number of people needed in a room to reach that threshold. For example, if you enter 25, the program will run and determine the number of people in a room needed for there to be a 25% chance that two people share the same birthday.